Split node_send() into two versions - one for calling when the node lock is not held, and a new one, node_send_locked(), for when it is. This allows messages to be sent from within event handling functions, such as worker_callback(), which previously had to do it "manually" by constructing the message object itself.
Both of these functions now take a sourcelocalid parameter instead of an endpoint.
This allows messages to be sent with a localid that does not necessarily correspond
to a real endpoint, which is of particular use for one-way messages where the source
does not actually have an endpoint, and the receiver does not need to reply. It is
also now permissible for a message to be sent to another endpoint on the same node.

Renamed node_connect() to node_connect_lock(), reflecting the fact that the node lock
should be held when this function is called, and fixed all callers so they do this.

Also fixed a race condition in the STARTTASK message handler in manager.c