Don't allocate frames (and their data area) individually using malloc(). This was causing signficiant memory fragmentation, causing the size of the process to grow well beyond its need (even though the total amount of memory handed out by malloc was small).
Instead, allocate large blocks of memory, divided up to hold the individual frames. The
stack data area of each frame is allocated as part of this, and sits directly at the end
of the main frame struct. This means that we can't resize the data area anymore, so we
have to use the largest stack size of all the functions in the program. This is a
tradeoff, but given that its only normally a small portion of memory used by the program
it seems to be worth it.