Made memory usage more efficient by avoiding calls to realloc() whever possible. When creating either an array* or carray*, we now allocate the required amount of memory from the start, if we know how much will be needed. The arrays can still grow dynamically if more elements are later appended, but in many cases (particularly IO), the array is just created with a fixed size and never grown.
Also changed the encoding of the IORESPONSE message so that the message fields are just
stored directly in the array of bytes that is sent, instead of marshalled/unmarshalled
by the functions in data.c. This is mainly to avoid extra memory copying when the
data is read from a socket and sent to the task through an IORESPONSE message.

array_new() now takes an extra argument containing the initial amount of entries to
allocate. If this is zero then a default count of 1024 is used.