Allow sparked frames to be added to the runnable queue even if there is only one task in the process
Properly handle import declarations in files that are themselves imported

Added modules util and http, in the modules directory. These are compiled into the
nreduce executable, to avoid having to keep track of them separately on the filesystem.

Added the built-in functions stringtonum1 and error1 (and their prelude.l wrappers
stringtonum and error, which ensure a fully eval'd string is passed)

Added reverse function