Fixed a very ugly bug that was causing partial evaluation to apply an if rule incorrectly.
Just goes to show that valgrind doesn't find all invalid memory accesses. In this case
the program was trying to access element -1 of a statically defined array. Thus the
behaviour of the partial evaluation would vary according to the way in which the
builtins.o file happened to get laid out by the compiler, and could therefore change
after changes were made to builtins.c in functions that weren't even called in some of
the tests.