Improvments to the module system. The convert_modules script is now use to take a set of ELC files and turn them into statically defined strings that are compiled into the nreduce executable. These are all placed in the one C file, and a table is generated containing the names and source of the modules. This table can be inspected by other parts of the code (e.g. the processing of the import statement), instead of having to hardcode tests for each of the modules. This makes adding new modules easier.
The prelude is now considered a module and integrated with this logic, but has the
special distinction that the functions in it are not given a prefix during parsing.

Rewote the profiling code so that more detailed output is produced, including a
printout of all source files in which each line is annotated with the number of
instructions executed that correspond to expressions on that line. The profiling
is activated when the engine is compiled with PROFILING defined. At task completion,
the output is written to PROFILE_FILENAME (currently set to "profile.out"). The
-s and -p  (--statistics and --profiling) flags have been removed. As before, profiling
is a compile time option for the engine since having it enabled slows down execution.