Optimised the way that lazy evaluation is done to minimise the number of MKFRAME instructions that are executed. Complex expressions such as those involving multiple function calls need to have several frames created, which costs extra time for the memory allocation, which is wasted if the frame never ends up being executed.
Instead, each complex expression that is in a non-strict context is now placed in a
separate supercombinator, so that only one MKFRAME is necessary. The expression that
comprises the body of the new supercombinator can be evaluated strictly when it is
activated. This significantly reduces the amount of memory allocation that occurs.

This lifting is done in a new compilation stage, referred to as "non-strict 
expression lifting". It must be done *after* strictness analysis, so the strictness
analysis stages is now called from within source_process() instead of
source_compile().

We also now avoid executing the GLOBSTART instruction, since this is a waste of time.
Even though it is a no-op, it still costs a small amount of time to do the instruction
pointer increment and switch statement. It is only there as a marker in the bytecode
to indicate the start of the new functon, and ideally it could eventually be removed
altogether.