Changed client mode so that it allows you to specify the list of nodes you want to start the process on. This list is contained in a file that must be supplied on the command line.
Modified the message sending primitive so that if an attempt is made to send a message
to a destination endpoint where we do not yet have a connection to that node, then a
connection to the node will be established. This simplifies the client code, in
particular relating to launching a process on multiple nodes, because all that is
necessary is to send the messages to the appropriate manager endpoints and the
connections will then be established.