Instead of finding out the local IP address based on the getsockname() result for the first opened network connection, use the result from the "hostname" command instead. This enables us to know what IP to use in endpoint identifiers right from the beginning, instead of having to wait until we have the first network connection.
This is necessary because the endpoint structure has now been changed to include
the full (ip,port,localid) identifier, for the convenience of threads that use
the messaging layer. Without knowing the IP right from the start, it would not
be possible to create endpoints (such as the manager) until the first network
connection was established.

The listen IP determined at startup is passed to the listen() call when the node
opens the listening port for connections from other nodes. This removes the
possibility that a connection from another node could be received on a local IP
that differs from the proper one. Thus, we no longer need to check that the
local address that another node has connected to is the same as what we expected.
If an attempt is made to connect to us on the same listening port but a different
address, then the connection will simply fail.

As a result of the endpoint id change, it is now simpler to send a message to a
local endpoint. Instead of manually having to construct the destination id, it
is possible to use the one obtained from the appropriate endpoint object.

Also changed from using struct in_addr to using in_addr_t to represent IP
addresses. The latter is typedef'd to uint32_t in <netinet/in.h>, and can thus
be used with the == operator, instead of memcmp().