Added a link feature similar to the one in erlang. An endpoint may set up a "link" to another endpoint (local or remote), in order to be notified of when it exits, or becomes inaccessible due to a network error. This can be used to handle fault tolerance - the messaging model we use is send and forget, so a thread that wants to deal with a remote endpoint's failure can do so by handling the ENDPOINT_EXIT message appropriately.
This mechanism is currently implemented only for the case when a connection to
another node drops. Any endpoints on the local node that had set up links with
endpoints on the remote node will receive an ENDPOINT_EXIT message for each link.
This replaces the logic in worker.c that would previously kill a task when it found
that any of the endpoints in its ID map were on the machine to which the connection
failed; now, the task itself handles the ENDPOINT_EXIT message, and causes itself to
exit.

During shutdown, the close routine for each of the thread types now sends a KILL
message to the apprioriate endpoint, instead of directly stopping the thread itself
(with the exception of the launcher, which still needs to be changed). This further
reduces the number of cases where we have to deal with shared state.

Added messages.h, which contains the structures specifying the format of different
messages. Currently this is only done for ENDPOINT_EXIT, but others will appear here
in the future. The MSG_* defines have also been moved here from runtime.h.