Changed the -d option to nreduce (which used to specify host:port) to -p (which now specifies just the port no). This is because we no longer provide the ability to listen on a particular IP address; nreduce always uses the one corresponding to the host name of the machine.
Added the ability to show timing information for chord tests. Useful for debugging
network performance when running against multiple nodes.

Made localids and chordids 32-bit unsigned integers. The first couple of localids are
reserved for the manager and chord threads, avoiding the chance of a duplicate
endpoint id being assigned to another thread which conflicts with these.

Avoid using memcmp for equality tests of endpointid objects. This is because the
change in localid size meant that the size of the endpointid struct increased such
that the fields did not use up the whole memory area, due to the way in which they
are aligned. Now we have to do the comparison on a field-by-field basis.

Added an option to use outgoing network connections only when sending messages.
This is enabled by setting the OUTGOING environment variable. Provides better
performance in cases where the tcp_abc option is set on some Linux kernels.