Enabled chord by default. Each node now starts a chord thread at startup, and uses this to get itself into a "grid" of nreduce instances running on other machines. The -i option to nreduce specifies an existing node to connect to.
Changed client mode to use chord to discover nodes to use for launching processes,
instead of requiring the user to specify the list of machines in a file. Now, the
client must be run as follows:

> nreduce --client <initial node> run <N> <program>

where N is the (maximum) number of machines across which the process can be used.

Additionally, if the following command is run:

> nreduce --client <initial node> findall

then the program will do a search for all machines in the grid and list them.

The run command in the console has also been removed; to run a program across multiple
nodes it is now necessary to do so using the client.