Changed all of the built-in network functions so that they interact with network connections indirectly by sending messages to the manager, rather than manipulating the node object directly. This is because it is necessary for us to be able to migrate sysobjects to other nodes when running in parallel, and a program may therefore try to read or write to a connection which actually lives on another machine. In such a case, a message will be sent to the manager on the remote machine and that will perform the actions necessary for the I/O operation.
A set of new messages has been added for listen, accept, connect, read, write, and their
*_response counterparts. These are exchanged between tasks and managers. The request
messages are sent to the manager from builtins.c, and the responses are handled in
interpreter.c. The manager has a new function added for each type of message to
handle that appropriately. No built-in functions have to lock the local node any more,
which also makes things a bit cleaner.

The network functions now also print more accurate error messages, since the errno
value of the failed operation is recorded and sent back in the response message, and
then passed to set_error().