Changed PNTR_VALUE to 0xFFF00000, which is a SNaN (Signaling NaN), instead of the old value 0xFFFC0000, which was a QNaN (Quet NaN). By using a signaling NaNs for representing pointers, we can get the FPU to do the work of detecting the case where a pointer is passed as an argument to an arithmetic operator, instead of having to do an additional software check first. While we still do the manual check at present, relying on floating point exceptions to detect such pointer usage will allow us to gain greater performance when the native code generator is working.
One implication of the use of SNaNs is that we cannot use 0 in the second for byte
to indicate a "null" pntr value, because this could be interpreted as negative infinity.
Instead, we store a 1 in the second byte. The is_nullpntr() macro has been updated to
check this. This is safe because the memory address 1 will never be returned by a normal
memory allocator. Also, the NULL_PNTR macro has been introduced which should be used
instead of make_pntr() when assigning a null pntr value.

The typedef for pntr is now a struct instead of a double. This is to avoid the compiler
generating floating point load/store instructions for a pntr assignment, which can cause
floating point exceptions if the bit pattern is a SNaN. It actually gives a slight
performance increase as well since a normal memory copy is used to transfer the data
instead.