Added a new opcode, CALL, which is similar to MKFRAME, but immediately executes the frame instead of pushing it onto the stack. This is used where it is known the frame needs to be evaluated straight away. It avoids allocation of a cell for the frame, and instead sets the new retf and retpos fields of the frame so that upon return, the stack of the calling frame will be updated directly. This is an alternative method of returning a value from a frame which can be used instead of having an associated cell which becomes an IND cell referencing the result.
This optimization is important for programs which use lots of function calls, because
it reduces the amount of heap usage, and avoids the overhead of sparking in certain
cases. In the nfib benchmark, it reduces the number of cell allocations (and thus
garbage collection cycles) by 50%.

CALL instructions are not added directly during compilation of a supercombinator; rather,
they are added by a peephole optimiser which has been added in bytecode.c. This catches
instances of SPARK/EVAL on a stack position and replaces it with EVAL (the spark is
unnecessary since that stack position will be evaluated straight away). Subsequently,
MKFRAME/EVAL pairs are caught and replaced with CALL/EVAL. The EVAL must remain because
the frame could return another frame which needs further evaluation. However in the
common case the EVAL will just see that the stack position contains a value, and will
not need to take further action.

The performance gains of this change are modest at present but are mainly useful for the
currently-in-progress native code generator.