Changed all xmlChar* references to char*. We're really not treating these properly, and the code will probably break if used with unicode text. But for current purposes treating them as C strings will suffice.
Properly handle identifiers that contain namespaces

Some optimisations to the XSLT support code to cut down the number of instructions
needed for simple operations such as determining if an item is atomic.

Optimised apmap3 and filter3 so that they don't need to keep around a reference
to the start of the list they are iterating over. This was causing very large
memory usage for long temporarily lists which were simply being produced and then
immediately consumed, and thus don't need to stay around in memory in their
entirety.