Workaround for a bug in the OS X kernel which sometimes misses delivery of a SIGTRAP signal. We were relying on this in the native execution engine because an interrupt to main program execution would be scheduled by writing the byte 0xCC (INT 3) to the appropriate places in memory so that the program would jump into the SIGTRAP handler at the end of the current bytecode instruction, which would then arrange for handle_interrupt() to be called.
Instead, we now write a CALL instruction to one or both breakpoint positions when we need to
interrupt execution. This has the drawback of requiring 5 bytes instead of 1, so the compiler
now has to ensure that where an instruction has two breakpoint addresses (currently only JFALSE
and EVAL), these are at least 5 bytes apart. It does this by padding with NOPs where necessary.

samples/mandserver.elc is the only program in the test suite which exposed the bug but I expect
it would probably occur with most network-intensive code.

As a result of this change, native code execution seems to work perfectly now on both OS X and
Linux.