Simplified the process of invoking the interrupt handler. Instead of going through through an extra stage via the code in Linterrupt, we just call handle_interrupt() directly from within handle_trap().
This change also avoids a race condition that was present in the old code, whereby an event
received after interrupt_again had been checked (and found to be false), but before the program
exited the Linterrupt code, the event could be missed. Now, SIGUSR1 just patches in the CALL
instruction to the eip saved on the stack even when it is in the trap handler.