Changed the way that nodes and atomic values are represented to use a consistent representation. Previously, different node types stored their properties in lists with each property in a different position, much like the way memory might be laid out in a class hierarchy where each subclass defined its own properties specific to that type.
Instead, we now represent all nodes as a list where the elements correspond to the properties
in the following order: type, value, parent, prev, next, nsuri, nsprefix, localname, attributes,
namespaces, children. The atomic types string, number, bool, and untyped use a 2-element list
containing just type and value. The atomic type qname uses the same representation as nodes, with
only the nsuri, nsprefix, and localname fields used.

This avoids the need for accessor functions for common properties which had to inspect the type
and access a different list position depending on what type. For example, to get the children
of an item, it was necessary to check if it was an element or document, and then call the
appropriate accessor function for that type. This, to put it mildly, sucked. So now you can just
call the function item_children to get the child elements of an item regardless of what type it
is.