Implemented support for sorting the result of path expressions by document order. This involved adding an extra element to each item which can be used to obtain the index of that item relative to the start of the document. These indices are compared to determine the relationship between two nodes in a result sequence during sorting.
Also added a reparent method, which makes a copy of a node tree, adjusting the parent, prev, and
next pointers appropriately. This is necessary when constructing elements within an XSLT program
(using either literal result elements or the xsl:element instruction), so that path expressions
which look at the nodes ancestors will work properly. Unfortunately it means a lot of extra
copying, but it should be able to reduce this in the future by setting the correct parent
pointers on creation of new elements wherever possible.

A few cases of document ordering are not yet covered, e.g. comparison of nodes from different
documents.