Added a new C-like syntax, which is a lot less painful to code in than ELC. Operations like appending strings and looping over a list are now a whole lot easier, and the use of infix operators feels much more natural. The built-in C indentation logic provided by emacs and other editors will also correctly handle this code, unlike ELC.
While this does carry the cost of a bit of extra complexity in the system, I believe it's worthit, both in terms of making my own job easier, and reducing the learning curve for other 
programmers. It actually doesn't require much changes, with basically the only modifications
being to the lexer and grammar files and a small number of extra support routines.

Both the old ELC syntax and the new one can be used side-by-side in the same file. Each function
is either written in the old syntax or the new syntax. The parser translates the new syntax into
the same internal representation, and it can be converted back into ELC using the routines that
are already in debug.c