Removed the tsk->ackmsg buffer that was written to as an address-containing message was parsed. This was originally used to send back a list of received in-flight addresses to the sender of the message, but for a long time the count was sent back instead and this buffer was never actually transmitted.
Simplified the error handling in decoding messages sent between tasks relating
to distributed heap management. Previously, each function would return a result
indicating whether or not the parsing of a field succeeded, in a similar manner to
PVM. While this could theoretically be used to make the VM more resillient against
corrupted data, it was uglier to deal with. The only way that invalid data could
be received is if a malicious client was deliberately sending the wrong things. For
simplicty at this stage we're going to assume all peers are trusted, and thus
invalid data means a bug, so we just abort. The previous code was doing this anyway,
just higher up in the call stack.

Changed the representation of seralised graph segments and simplified the process
of sending references vs. pointers. This includes correct handling of arrays, and
recording a received object in the global address table so it doesn't have to be
sent the next time that address is needed. This makes distributed execution much
more stable, since the crashes that were occurring previously were mostly related
to arrays.

Got rid of endpt->interruptptr, which used to point to the stack-local variable
in interpreter_thread() that would get set whenever a message arrived. This had to
be done to avoid a race condition which occurred if a task received an interrupt
just after interpreter_thread() had returned but the endpoint had not yet been
cleaned up. Instead, we now use the endpt->interrupt field. Theortically this is
more expensive to check on each instruction cycle, but tests showed the performance
impact to be neglible (less than 1%), and anyway it only affects interpreted mode
(native code still receives interrupts via signals).

Changed most of the aborts()s into calls to fatal(), which is much more useful in
the absence of a working debugger or when remote nodes have their output redirected
to a log file.

Made distributed processes exit properly when finished. This involves a call to the
new built-in exit() function at the end of the program, which causes the calling
task to exit, and the others too once they receive the ENDPOINT_EXIT message. This
is hidden from the programmer by doing it within the __start wrapper function in
prelude.elc.

Added a configure.in check for execinfo.h, so backtraces work again on linux.

Fixed a few other minor bugs and crashes.