Cleanups to the communications infrastructure:
Got rid of the generic event dispatching mechanism in node. Basically the only thing
it was used for was the worker_callback() function which dealt with threads that
wanted to use "regular" socket connections, and also a couple of other minor places
like the client detecting connection failure. I decided there was unnecessary
complexity here, and so replaced the worker_callback() function with a set of
notify_* functions which would do the job of sending messages relating to
connections.

This change has enabled some functionality which previously needed to access
private node functions to be moved entirely within node.c, so that these functions
no longer need to be exposed to other files. An example of this is
node_send_locked(). As much as possible different threads should interact only
through message passing, as this reduces the chances of race conditions, and
is the same model as is used for distribution.

Renamed node_send() and endpoint_next_message() to endpoint_send() and
endpoint_receive(), respectively.

Fixed an ugly obscure bug in array_vprintf() which would sometimes not append the
last character.

Changed the console to use connections in the same way that tasks do. This removes
the concept of there being special "console" connections (the now have isreg=1),
and avoids the need for the console to directly access private data structures and
functions of the node.

Made error handling work correctly for the case where output is being sent to a
remote client.