Made it not an error to call endpoint_link() where the specified destination already exists in the outgoing links set. In this case it is simply not added again.
In the connection listing in the debug console, show the owner (if any) of each connection.

Establish a link between the manager and any endpoint which opens a connection or listener. This
is so that if the endpoint exits without manually deleting the connection or listener, then the
cleanup can still occur. This also allows console connections to be deleted automatically when
the console thread exits.

If any tasks exit unexpectedly (e.g. due to a kill command from the debug console), have the
client's output thread print an error message and exit.