Added a foreign function interface that enables ELC code to make calls to through to java.
This is achieved by running a java "bridge" which acts as a server. nreduce connects to this
and sends commands using a simple text-based protocol to create objects and invoke methods.
The jnew and jcall functions can be used to communicate with java code. Primitive types and
java object references can be sent back and forth as parameters and return values.

To use the java functionality, the java bridge must be running on port 2001:
# cd java
# java nreduce.Bridge 2001

See samples/java.elc and tests/java/* for examples. samples/graphics.elc gives a nice
demo of how to use the java swing APIs to visualise the results of a computation performed
in ELC.