Moved all private fields of the node and endpoint structures into private structures that are declared in netprivate.h. This prevents them from being used by code in other parts of the tree and makes the public interface of the networking library cleaner.
Separated out the contents of messages.h into parts that are specific to nreduce, and generic
messages that are part of the networking library and could be used by any application. The former
remain in messages.h which is now in the runtime directory, while the latter have been moved
to node.h.

Changed the message numbering so that the ones used by the networking library use values close
to 2^31-1, and application-specific messages begin from 0.

Other minor cleanup/changes
- removed print_endpointid() and changed all code that used it to now use EPID_FORMAT and EPID_ARGS
- moved the contents of netutil.c into util.c
- renamed messages.c to send.c
- removed the node parameter from lookup_hostname() and lookup_address()