Added a "reason" field to endpoint_exit_msg. It is set to 1 on normal exit, and 0 if an error occurs.
Removed some calls to node_log() that weren't necessary. node_connect_locked() and
node_listen_locked() now require the errmsg parameter to be supplied, and always record errors
in that instead of writing a log message.

Properly handle a connection error in node_send_locked where the error happens immediately
(i.e. not when the result of the asynchronous call later becomes available)