Added the spawn function, which creates a child process and returns a stream containing its output. Spawn also takes a parameter specifying the input to the process, so a sort of bi-directional pipe exists between the parent and child.
This works by using the socketpair() fuction to create a pair of sockets which are used just like
normal TCP connections. The existing mechanisms for sending and receiving data over a connection
are used for interaction.

Also added a compile function which can be called from user code to generate bytecode for a given
ELC program. This must be called before invoking spawn, since spawn takes the bytecode to launch
the new process with rather than the source code.

The spawn function can be used to implement a CGI-like mechanism for a server process, where each
request is handled by a separate process that is launched from the server.