Added optimisations for numeric expressions.
A type checking phase is performed to traverse the syntax tree and assign a type to each node. The
main two types are NUMBER and GENERAL (i.e. not number). Those marked with NUMBER are compiled
in such a way that they avoid the boxing/unboxing of values between different parts of the
expression evaluation. This dramatically improves performance of numerically intensive code.

In some cases, additional copies of a function is made, in case it is called both with and without
numeric arguments. Up to two copies (derivatives) of each function can be made, which are more
efficient but rely on numeric parameters being passed.