Added support for service choreography, where data is transferred directly between service hosts rather than via some other host. This requires VM nodes to be deployed on all hosts on which services to be choreographed are running.
This is achieved by making any function that access a sysobject migrate to the host on which
the "master" version of that sysobject resides. Currently the only functions which do this
are readcon and write_data. Additionally, _connect will migrate to the destination host if
the process contains a task that is on that host.

To support migration of built-in functions, the bytecode compiler now enforces that all
impure built-ins must be called via their wrapper functions, defined in bytecode. The reason
for this is that when a function migrates, it restarts from the beginning, so that it can
execute the appropriate EVAL instructions to obtain the necessary local copies of the cells
the function needs. This migration is only safe to do for the built-in function wrappers.

The global address table logic has also been modified to distinguish between target addresses
(the location of the master copy of an object, for remote references or replicas) and
physical addresses (the address of a ref or replica in the local heap). I realised I wasn't
handling this correctly previously, which meant that object distribution didn't work
properly in certain scenarios. Significant modifications/rewrites have been made to all the
major functions that use the GAT.