Added an experimental strict evaluation mode, in which all supercombinator arguments are assumed to be strict unless otherwise specified (by prefixing them with @). This evaluation mode will work with almost all programs that work with lazy evaluation, with the exception of those that explicitly rely on things that lazy evaluation provides (e.g. support for infinite data structures).
The reason for adding this evaluation mode is to enable "true" automatic parallelisation for
situations in which the programmer would otherwise have to add lots of annotations throughout
a program in order to trigger sparking.

Lazy evaluation is still the default at this stage until strict evaluation gets more testing.
Strict evaluation can be turned on by setting the environment variable STRICT to any value.

Also fixed a small bug in the native code generation for the ITEMN opcode.