Restructured the connection management code as a finite state machine. This makes it a lot easier to understand what the different paths are through a connection's lifecycle, and what events/transitions can happen in which states.
This is part of an effort to have clients "control themselves" better when sending out service
requests. There is now a limit on the number of connections that can be in an "opening" phase,
i.e. those for which the client has not obtained confirmation of acceptance by the server.
Since TCP does not send anything to the client to indicate when the server process has called
accept(), the best we can do in this case is base the confirmation on whether or not we have
read at least 1 byte of data over the connection (meaning that it must have been accepted).

The diagram notes/connection_fsm.graffle shows the FSM. This is implemented in connection.c,
with almost everything that happens with a connection (with the exception of reading and
writing) being controlled by "sending" events to the FSM by calling connection_fsm().