When using strict evaluation, have b_readcon() start the next frame running to continue reading the stream immediately. This is needed to force the complete input stream to be evaluated as soon as possible.
Modified Server.java to make sure that it reads all of the data from the client, and consumes
the FIN, before closing the socket. Not doing this causes connections to be aborted, resulting
in the client missing data or getting errors (see next change).

Modified the load balancer so that if an error occurs reading from one end of a connection
pair, then do an abortive close of the connections rather than a normal one. This will
indicate to the other end that there was actually an error, rather than a normal close. This
was causing clients to miss the above error and just return truncated data when they should
have actually been reporting a "connection reset by peer" error.