Got rid of the UPDATEREF message, which was causing race conditions, because it was possible for a frame's result value to be sent back to the reference before it's scheduled address (if the frame had been migrated).
Fixed a bug in the garbage collector which was freeing target globals for
objects which still existed in the heap. This was causing replica sysobjects
to lose their association with their canonical versions, meaning that they
would appear as new objects when subsequently sent to another task.

With strict evaluation disabled, the pipeline test for choreography now works.