Changed the way that the amount of local service requests is limited. Instead of not sparking any frames at all when there are requests in progress, allow frames to be sparked, *except* when those frames correspond to connection attempts that were previously denied by b_connect because it noticed there were already enough requests in progress.
When b_connect rejects a connection request on this basis, it places it back into the SPARKED
state, and marks it as "nolocal". This is an indication to handle_interrupt that it should not
spark the frame directly if there are already enough requests in progress (as defined by
MAX_SVCBUSY). The frame can still be exported to other machines however, since if they send
a FISH request it means they want more work.

Also fixed some bugs in the way that frame migration is handled. Previously, the source node
on which a frame originated would replace the frame with a remoteref to (-1,-1). This caused
problems when the frame was migrated back to its home site and then away again, because each
time this happened, it would get a different global object with (-1,-1) associated with it.
Now, when a frame is sent from its home site, the remoteref points to the global associated
with the frame's physical address. This means that the reference isn't actually local, but
this is ok, since fetching is set to 1, and thus the address is never used to send out FETCH
requests.