Added options to set the number of frames that are sent out in response to a FISH request, and enable or disable postponing of connections. These are set via the environment variables OPT_FISHFRAMES and OPT_POSTPONE.
Changed the nolocal field of the frame struct to postponed.

Make use of the doubly-linked list of sparks when responding to a FISH request,
instead of doing a heap traversal.

Before sending a cons cell or array across the network, try to expand it first,
in order to try and avoid a future FETCH request for the next portion of the
list.

Changed the distributed garbage collection interval from 20 to 30 seconds.

Made the built-in arrayprefix function (used by append) strict in all arguments
when the VM is running in strict mode.

When expanding arrays, try to expand strings as well.